requirements.txt
streamlit
pandas
requests
matplotlib
Pillow
python-dotenv
cryptocmd
